requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
gunicorn==22.0.0
python-dotenv==1.0.1
bcrypt==4.2.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
jinja2==3.1.4
httpx==0.27.2
supabase==2.9.1
user-agents==2.2.0
itsdangerous==2.2.0
aiofiles==24.1.0
